Move unit tests for Metric and MetricSet to the 'metrics' module.
