+++ metrics/src/test/java/ai/vespa/metrics/MetricSetTest.java
@@ -1,8 +1,6 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.model.admin.monitoring;
+package ai.vespa.metrics;
 
-import ai.vespa.metrics.Unit;
-import ai.vespa.metrics.VespaMetrics;
 import ai.vespa.metrics.set.Metric;
 import ai.vespa.metrics.set.MetricSet;
 import com.google.common.collect.Sets;
